ISSUE #315 * Create `minos.common.snapshot` as module.
