+++ minos/common/snapshot/__init__.py
@@ -1,0 +1,13 @@
+"""
+Copyright (C) 2021 Clariteia SL
+
+This file is part of minos framework.
+
+Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
+"""
+from .abc import (
+    MinosSnapshot,
+)
+from .memory import (
+    InMemorySnapshot,
+)
