+++ tests/test_common/test_model/test_aggregate/test_base.py
@@ -8,7 +8,7 @@ Minos framework can not be copied and/or distributed without the express permiss
 import unittest
 
 from minos.common import (
-    InMemoryMinosSnapshot,
+    InMemorySnapshot,
     MinosInMemoryRepository,
     MinosRepositoryAggregateNotFoundException,
     MinosRepositoryDeletedAggregateException,
@@ -25,12 +25,12 @@ from tests.utils import (
 
 class TestAggregate(unittest.IsolatedAsyncioTestCase):
     async def test_create(self):
-        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemoryMinosSnapshot() as snapshot:
+        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemorySnapshot() as snapshot:
             car = await Car.create(doors=3, color="blue", _broker=broker, _repository=repository, _snapshot=snapshot)
             self.assertEqual(Car(1, 1, 3, "blue", _broker=broker, _repository=repository, _snapshot=snapshot), car)
 
     async def test_create_raises(self):
-        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemoryMinosSnapshot() as snapshot:
+        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemorySnapshot() as snapshot:
             with self.assertRaises(MinosRepositoryManuallySetAggregateIdException):
                 await Car.create(
                     id=1, doors=3, color="blue", _broker=broker, _repository=repository, _snapshot=snapshot
@@ -44,7 +44,7 @@ class TestAggregate(unittest.IsolatedAsyncioTestCase):
         self.assertEqual("tests.aggregate_classes.Car", Car.classname)
 
     async def test_get(self):
-        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemoryMinosSnapshot() as snapshot:
+        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemorySnapshot() as snapshot:
             originals = [
                 await Car.create(doors=3, color="blue", _broker=broker, _repository=repository, _snapshot=snapshot),
                 await Car.create(doors=3, color="red", _broker=broker, _repository=repository, _snapshot=snapshot),
@@ -56,7 +56,7 @@ class TestAggregate(unittest.IsolatedAsyncioTestCase):
             self.assertEqual(originals, recovered)
 
     async def test_get_one(self):
-        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemoryMinosSnapshot() as snapshot:
+        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemorySnapshot() as snapshot:
             original = await Car.create(
                 doors=3, color="blue", _broker=broker, _repository=repository, _snapshot=snapshot
             )
@@ -65,12 +65,12 @@ class TestAggregate(unittest.IsolatedAsyncioTestCase):
             self.assertEqual(original, recovered)
 
     async def test_get_one_raises(self):
-        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemoryMinosSnapshot() as snapshot:
+        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemorySnapshot() as snapshot:
             with self.assertRaises(MinosRepositoryAggregateNotFoundException):
                 await Car.get_one(0, _broker=broker, _repository=repository, _snapshot=snapshot)
 
     async def test_update(self):
-        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemoryMinosSnapshot() as snapshot:
+        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemorySnapshot() as snapshot:
             car = await Car.create(doors=3, color="blue", _broker=broker, _repository=repository, _snapshot=snapshot)
 
             await car.update(color="red")
@@ -82,12 +82,12 @@ class TestAggregate(unittest.IsolatedAsyncioTestCase):
             self.assertEqual(car, await Car.get_one(car.id, _broker=broker, _repository=repository, _snapshot=snapshot))
 
     async def test_update_raises(self):
-        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemoryMinosSnapshot() as snapshot:
+        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemorySnapshot() as snapshot:
             with self.assertRaises(MinosRepositoryManuallySetAggregateVersionException):
                 await Car(1, 1, 3, "blue", _broker=broker, _repository=repository, _snapshot=snapshot).update(version=1)
 
     async def test_refresh(self):
-        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemoryMinosSnapshot() as snapshot:
+        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemorySnapshot() as snapshot:
             car = await Car.create(doors=3, color="blue", _broker=broker, _repository=repository, _snapshot=snapshot)
             car2 = await Car.get_one(car.id, _broker=broker, _repository=repository, _snapshot=snapshot)
             await car2.update(color="red", _broker=broker, _repository=repository, _snapshot=snapshot)
@@ -98,7 +98,7 @@ class TestAggregate(unittest.IsolatedAsyncioTestCase):
             self.assertEqual(Car(1, 3, 5, "red", _broker=broker, _repository=repository, _snapshot=snapshot), car)
 
     async def test_delete(self):
-        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemoryMinosSnapshot() as snapshot:
+        async with FakeBroker() as broker, MinosInMemoryRepository() as repository, InMemorySnapshot() as snapshot:
             car = await Car.create(doors=3, color="blue", _broker=broker, _repository=repository, _snapshot=snapshot)
             await car.delete()
             with self.assertRaises(MinosRepositoryDeletedAggregateException):
