+++ tests/test_common/test_model/test_aggregate/test_with_posgres.py
@@ -14,7 +14,7 @@ from dependency_injector import (
 )
 
 from minos.common import (
-    InMemoryMinosSnapshot,
+    InMemorySnapshot,
     MinosRepositoryDeletedAggregateException,
     PostgreSqlMinosRepository,
 )
@@ -38,7 +38,7 @@ class TestAggregateWithPostgres(PostgresAsyncTestCase):
         self.container = containers.DynamicContainer()
         self.container.event_broker = providers.Object(FakeBroker())
         self.container.repository = providers.Object(PostgreSqlMinosRepository.from_config(config=self.config))
-        self.container.snapshot = providers.Object(InMemoryMinosSnapshot())
+        self.container.snapshot = providers.Object(InMemorySnapshot())
         await self.container.repository().setup()
         self.container.wire(modules=[sys.modules[__name__]])
