+++ tests/test_common/test_repository/test_entries.py
@@ -11,7 +11,7 @@ from datetime import (
 )
 
 from minos.common import (
-    InMemoryMinosSnapshot,
+    InMemorySnapshot,
     MinosRepositoryAction,
     MinosRepositoryEntry,
     MinosRepositoryUnknownActionException,
@@ -66,7 +66,7 @@ class TestMinosRepositoryEntry(unittest.IsolatedAsyncioTestCase):
         self.assertEqual(datetime(2020, 10, 13, 8, 45, 32), entry.created_at)
 
     async def test_from_aggregate(self):
-        async with FakeBroker() as broker, FakeRepository() as repository, InMemoryMinosSnapshot() as snapshot:
+        async with FakeBroker() as broker, FakeRepository() as repository, InMemorySnapshot() as snapshot:
             car = Car(1, 1, 3, "blue", _broker=broker, _repository=repository, _snapshot=snapshot)
             entry = MinosRepositoryEntry.from_aggregate(car)
         self.assertEqual(car.id, entry.aggregate_id)
